% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.ph_abort}
\alias{.ph_abort}
\title{Internal helper: .ph_abort}
\usage{
.ph_abort(headline, step = NULL, bullets = NULL, ...)
}
\description{
Emit an ERROR log block and abort execution.
}
\keyword{internal}
