% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{.ph_attach_finalizer}
\alias{.ph_attach_finalizer}
\title{Attach an auto-finalizer for phip_data connections}
\usage{
.ph_attach_finalizer(x)
}
\arguments{
\item{x}{A valid \code{phip_data} object.}
}
\value{
A \code{phip_data} object with an attached finalizer environment.
}
\description{
Creates a small environment that stores the current connection
handles and registers a finalizer to close them when the object is GC'd.
}
\keyword{internal}
