% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.ph_check_cond}
\alias{.ph_check_cond}
\title{Internal helper: .ph_check_cond}
\usage{
.ph_check_cond(
  condition,
  error_message,
  error = TRUE,
  step = NULL,
  bullets = NULL,
  ...
)
}
\description{
Execute an expression with start/stop timing logs.
}
\keyword{internal}
