% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.ph_check_extension}
\alias{.ph_check_extension}
\title{Internal helper: .ph_check_extension}
\usage{
.ph_check_extension(name, x_name, ext_vec)
}
\description{
Validate a filename extension against an allowed set.
}
\keyword{internal}
