% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.ph_check_null_default}
\alias{.ph_check_null_default}
\title{Internal helper: .ph_check_null_default}
\usage{
.ph_check_null_default(x, x_name, method, default)
}
\description{
Replace NULL with a default, logging a warning.
}
\keyword{internal}
