% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.ph_check_path}
\alias{.ph_check_path}
\title{Internal helper: .ph_check_path}
\usage{
.ph_check_path(path, arg_name, extension, is_dir = FALSE)
}
\description{
Validate a file or directory path (and optional extension).
}
\keyword{internal}
