% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-methods.R
\name{.ph_extract_data_long}
\alias{.ph_extract_data_long}
\title{Internal helper: .ph_extract_data_long}
\usage{
.ph_extract_data_long(obj)
}
\description{
Return the \code{data_long} slot if input is phip_data; otherwise
return the input unchanged.
}
\keyword{internal}
