% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.ph_log_info}
\alias{.ph_log_info}
\title{Internal helper: .ph_log_info}
\usage{
.ph_log_info(
  headline,
  step = NULL,
  bullets = NULL,
  verbose = .ph_opt("verbose", TRUE)
)
}
\description{
Emit an INFO log block if verbose logging is enabled.
}
\keyword{internal}
