% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.ph_log_ok}
\alias{.ph_log_ok}
\title{Internal helper: .ph_log_ok}
\usage{
.ph_log_ok(
  headline,
  step = NULL,
  bullets = NULL,
  verbose = .ph_opt("verbose", TRUE)
)
}
\description{
Emit an OK log block if verbose logging is enabled.
}
\keyword{internal}
