% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-methods.R
\name{.ph_modify_pd}
\alias{.ph_modify_pd}
\title{Internal helper: .ph_modify_pd}
\usage{
.ph_modify_pd(.data, new_counts)
}
\description{
Replace \code{data_long} in a phip_data object and return it.
}
\keyword{internal}
