% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{.ph_refresh_finalizer}
\alias{.ph_refresh_finalizer}
\title{Refresh finalizer connection list}
\usage{
.ph_refresh_finalizer(x)
}
\arguments{
\item{x}{A valid \code{phip_data} object.}
}
\value{
A \code{phip_data} object with refreshed finalizer connections.
}
\description{
Updates the finalizer environment with the current connections.
}
\keyword{internal}
