% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{.ph_sync_peptide_con}
\alias{.ph_sync_peptide_con}
\title{Sync peptide connection handle}
\usage{
.ph_sync_peptide_con(x)
}
\arguments{
\item{x}{A valid \code{phip_data} object.}
}
\value{
A \code{phip_data} object with \code{meta$peptide_con} updated.
}
\description{
Ensures \code{meta$peptide_con} mirrors the peptide library
attribute.
}
\keyword{internal}
