% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.ph_with_timing}
\alias{.ph_with_timing}
\title{Internal helper: .ph_with_timing}
\usage{
.ph_with_timing(
  headline,
  step = NULL,
  bullets = NULL,
  expr,
  verbose = .ph_opt("verbose", TRUE)
)
}
\description{
Conditionally raise a formatted warning or error.
}
\keyword{internal}
