% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.ph_word_list}
\alias{.ph_word_list}
\title{Internal helper: .ph_word_list}
\usage{
.ph_word_list(word_list = NULL, and_or = "and", is_are = FALSE, quotes = FALSE)
}
\description{
Build a human-readable list from a character vector, with
optional quoting and conjunction handling.
}
\keyword{internal}
