% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-methods.R
\name{export_parquet}
\alias{export_parquet}
\title{Export a phip_data Table to Parquet}
\usage{
export_parquet(x, path)
}
\arguments{
\item{x}{A <phip_data> object.}

\item{path}{File path (character) to save the output \code{.parquet} file.}
}
\value{
NULL (invisibly).
}
\description{
Exports the \code{data_long} table from a \strong{phip_data} object to disk in Apache
Parquet format.
}
\note{
The export is performed directly and efficiently from the
database/lazy table without reading all data into memory.
}
\examples{
pd <- load_example_data()
out_path <- tempfile(fileext = ".parquet")
export_parquet(pd, out_path)
unlink(out_path)

}
