% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-methods.R
\name{merge.phip_data}
\alias{merge.phip_data}
\title{Merge or join a \code{phip_data} object}
\usage{
\method{merge}{phip_data}(x, y, ...)
}
\arguments{
\item{x}{A \code{phip_data} object.}

\item{y}{A data-frame-like object \emph{or} another \code{phip_data}.}

\item{...}{Arguments forwarded to either \code{\link[base:merge]{base::merge()}} or the chosen
\strong{dplyr} join (e.g. \verb{by =}, \verb{suffix =}, etc.).}
}
\value{
A new \code{phip_data} whose \code{data_long} contains the merged / joined
tibble.
}
\description{
Merge or join a \code{phip_data} object
}
