% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{phip_example_path}
\alias{phip_example_path}
\title{Path to example PhIP-Seq datasets shipped with phiperio}
\usage{
phip_example_path(name = c("phip_mixture"))
}
\arguments{
\item{name}{Character scalar. Name of the example dataset.
Currently supported: \code{"phip_mixture"}.}
}
\value{
A character scalar with an absolute path to the file.
}
\description{
Path to example PhIP-Seq datasets shipped with phiperio
}
\examples{
sim_path <- phip_example_path("phip_mixture")
# phip_obj <- phip_convert(sim_path)

}
