% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-validate.R
\name{validate_phip_data}
\alias{validate_phip_data}
\title{Internal validator for <phip_data> objects}
\usage{
validate_phip_data(x, na_warn_thresh = 0.5, auto_expand = TRUE)
}
\description{
Internal validator for <phip_data> objects
}
\keyword{internal}
