% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{phiperio-package}
\alias{phiperio}
\alias{phiperio-package}
\title{phiperio}
\description{
Provides utilities to import, validate, and manage PhIP-Seq datasets, including standardized conversion pipelines, data checks, and access to cached peptide metadata.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Polymerase3/phiperio}
  \item \url{https://polymerase3.github.io/phiperio/}
  \item Report bugs at \url{https://github.com/Polymerase3/phiperio/issues}
}

}
\author{
\strong{Maintainer}: Mateusz Kolek \email{mati.kolek13@gmail.com} (\href{https://orcid.org/0000-0001-6470-4830}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Alon Alexander \email{a.alexander@umcg.nl} [contributor, copyright holder]
}

}
\keyword{internal}
